Add files via upload
This is a Ducky Script created to help with Microsoft Intue enrollment. It creates a dir in the C drive called HWID then installs the AutopilotInfo script and  creates the CSV file containing the HWID required when enrolling devices to MS Intune.  This script also copys the file to a usb mounted as D:\ in a path called "D:\HWID\" for ease.